requirements.txt
rank-bm25==0.2.2
tqdm==4.66.5
python-dotenv==1.0.1
langchain==0.3.3
langchain-openai==0.2.2
pypdf==5.0.1